% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm.R
\name{collinearity}
\alias{collinearity}
\title{collinearity() returns sample collinearity on 0 to 1 scale as function of generalized correlation}
\usage{
collinearity(X)
}
\description{
collinearity() returns sample collinearity on 0 to 1 scale as function of generalized correlation
}
