% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm.R
\name{combnDeg}
\alias{combnDeg}
\title{plm: get all polynomial degrees terms of given predictor variables
X1, X2, deg=2 -> X1, X2, X1^2, X2^2, X1*X2
combnDeg: distribute (deg) degrees to (n) different X's
deg_plm: deal with nondummy terms
only_dummy: deal with dummy terms
n = number of X variables (eg. X1, X2, X3 -> n=3)}
\usage{
combnDeg(n, deg)
}
\description{
plm: get all polynomial degrees terms of given predictor variables
X1, X2, deg=2 -> X1, X2, X1^2, X2^2, X1*X2
combnDeg: distribute (deg) degrees to (n) different X's
deg_plm: deal with nondummy terms
only_dummy: deal with dummy terms
n = number of X variables (eg. X1, X2, X3 -> n=3)
}
