\name{currency}
\alias{currency}

\title{
Pre-Euro Era Currency Fluctuations
}

\description{
From Wai Mun Fong and Sam Ouliaris, "Spectral Tests of the Martingale       
Hypothesis for Exchange Rates", Journal of Applied Econometrics, Vol.
10, No. 3, 1995, pp. 255-271.  Weekly exchange rates against US dollar,
over the period 7 August 1974 to 29 March 1989.        
}

