\name{day,day1}
\alias{day}
\alias{day1}
\alias{day2}

\title{
Bike sharing data.
}

\description{
This is the Bike Sharing dataset (day records only) from the UC Irvine
Machine Learning Dataset Repository.  Included here with 
permission of Dr. Hadi Fanaee.  

The \code{day} data is as on UCI; \code{day1} is modified so that the
numeric weather variables are on their original scale.

The \code{day2} is the same as \code{day1}, except that \code{dteday}
has been removed, and \code{season}, \code{mnth}, \code{weekday} and
\code{weathersit} have been converted to R factors.

See \url{https://archive.ics.uci.edu/ml/datasets/bike+sharing+dataset}
for details.
}


