\name{english}
\alias{english}

\title{
English vocabulary data
}

\description{

The Stanford WordBank data on vocabulary acquisition in young children.
The file consists of about 5500 rows. (There are many NA values, though,
and only about 2800 complete cases.) Variables are age, birth order,
sex, mother's education and vocabulary size.

}

