% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyreg_sparklyr.R
\name{fit_sparklyr}
\alias{fit_sparklyr}
\title{Generalized Linear Model for Polynomial Data}
\usage{
fit_sparklyr(df = gen_df_sparklyr())
}
\arguments{
\item{df}{sparklyr DataFrame with response variable column named, 'y'}
}
\description{
Generalized Linear Model for Polynomial Data
}
