% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfFSR.R, R/FSR.R
\name{FSR}
\alias{FSR}
\alias{FSR}
\title{FSR}
\usage{
FSR(Xy, max_poly_degree = 4, max_interaction_degree = 2,
  cor_type = "kendall", threshold = -1.001, pTraining = 0.8,
  pValidation = 0.2, max_block_size = 250, noisy = TRUE, seed = NULL,
  model = "lm")

FSR(Xy, max_poly_degree = 4, max_interaction_degree = 2,
  cor_type = "kendall", threshold = -1.001, pTraining = 0.8,
  pValidation = 0.2, max_block_size = 250, noisy = TRUE, seed = NULL,
  model = "lm")
}
\arguments{
\item{Xy}{matrix or data.frame; outcome must be in final column.}

\item{max_poly_degree}{highest power to raise continuous features; default 4.}

\item{max_interaction_degree}{highest interaction order; default 2. Also interacts each level of factors with continuous features.}

\item{cor_type}{correlation to be used for pseudo R^2. Default Kendall's (robust).}

\item{threshold}{minimum improvement to keep estimating.}

\item{pTraining}{portion of data for training}

\item{pValidation}{portion of data for validation}

\item{max_block_size}{Most of the linear algebra is done recursively in blocks to ease memory managment. Default 250. Changing up or down may slow things...}

\item{noisy}{display measures of fit, progress, etc. Recommended.}

\item{seed}{Automatically set but can also be passed as paramater.}

\item{pTesting}{portion of data for final test}

\item{Xy}{matrix or data.frame; outcome must be in final column.}

\item{max_poly_degree}{highest power to raise continuous features; default 4.}

\item{max_interaction_degree}{highest interaction order; default 2. Also interacts each level of factors with continuous features.}

\item{cor_type}{correlation to be used for pseudo R^2. Default Kendall's (robust).}

\item{threshold}{minimum improvement to keep estimating (pseudo R^2 so scale 0 to 1). Default -1.001 means 'estimate all'.}

\item{pTraining}{portion of data for training}

\item{pValidation}{portion of data for validation}

\item{max_block_size}{Most of the linear algebra is done recursively in blocks to ease memory managment. Default 250. Changing up or down may slow things...}

\item{noisy}{display measures of fit, progress, etc. Recommended.}

\item{seed}{Automatically set but can also be passed as paramater.}
}
\value{
list with slope coefficients, model details, and fit vs. the test data of the final model

list with slope coefficients, model details, and measures of fit
}
\description{
FSR

FSR
}
