% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyreg_r.R
\name{gen_dat_r}
\alias{gen_dat_r}
\title{Generate Data as DataFrame for Polynomial Regression}
\usage{
gen_dat_r(B = c(4, 6, 2, -1), n = 10000, rng = c(-2, 4), err = c(0, 4))
}
\arguments{
\item{B}{polynoimal coefficients, a number vector of length two or greater}

\item{n}{number of samples to generate}

\item{rng}{min and max of range to generate from as number vector length two}

\item{err}{mean and variance for normal distribution as noise to add to data set}
}
\description{
Generate Data as DataFrame for Polynomial Regression
}
\examples{
dat <- gen_dat_r()
dat <- gen_dat_r(B=c(5,4,3,-2),n=4000,rng=c(-3,5),err=c(0,5))
}
