% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyreg_sparklyr.R
\name{get_sc}
\alias{get_sc}
\title{Get sc object if one exists}
\usage{
get_sc()
}
\description{
If a variable sc exists, it is returned. If not, one is created using 
sparklyr::sparkconfig setting master to 'yarn-client'.
}
