% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyreg_r.R
\name{glm_r}
\alias{glm_r}
\title{Generalized Linear Model for Polynomial Data}
\usage{
glm_r(dat = gen_dat_r())
}
\arguments{
\item{dat}{DataFrame with response variable column named, 'y'}
}
\description{
Generalized Linear Model for Polynomial Data
}
