\name{mlb}
\alias{mlb}

\title{
Major Leage Baseball player data set.
}

\description{
Heights, weights, ages etc. of major league baseball players.  A new
variable has been added, consolidating positions into Infielders,
Outfielders, Catchers and Pitchers.

Included here with the permission of the UCLA Statistics Department.
}

