\name{mlens}
\alias{mlens}

\title{
MovieLens User Summary Data
}

\description{
The MovieLens dataset, \url{http://grouplens.org/},
is a standard example in the recommender systems literature.  Here we
give demographic data for each user, plus the mean rating and number of
ratings.  One may explore, for instance, the relation between ratings
and age.
}

