\name{newadult}
\alias{newadult}
\alias{newAdult}

\title{
UCI adult income data set, adapted
}

\description{
This data set is adapted from
the Adult data from the UCI Machine Learning Repository,
which was in turn adapted from Census data on adult incomes and other 
demographic variables.  The UCI data is used here with permission 
from Ronny Kohavi.

The variables are:

\itemize{
   \item \code{gt50}, which converts the original \code{>50K} variable
   to an indicator variable; 1 for income greater than $50,000, else 0

   \item \code{edu}, which converts a set of education levels to
   approximate number of years of schooling

   \item \code{age}

   \item \code{gender}, 1 for male, 0 for female

   \item \code{mar}, 1 for married, 0 for single
}

Note that the education variable is now numeric.

}

