\name{oliveoils}
\alias{oliveoils}

\title{
Italian olive oils data set.
}

\description{
Italian olive oils data set, as used in \emph{Graphics of Large 
Datasets: Visualizing a  Million}, by  Antony Unwin, Martin Theus and 
Heike Hofmann, Springer, 2006.  Included here with permission of Dr. 
Martin Theus.  
}

