% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyreg_r.R
\name{plot_r}
\alias{plot_r}
\title{Plot Data and Fitted Model using stats::glm}
\usage{
plot_r(dat = gen_dat_r(), lm = NULL)
}
\arguments{
\item{dat}{Polynomial Data}

\item{lm}{Linear Regression Fit on Polynomial Data}
}
\description{
Plot Data and Fitted Model using stats::glm
}
\examples{
dat <- gen_dat_r(B=c(5,4,3,-2),n=4000,rng=c(-3,5),err=c(0,5))
lm  <- glm_r(dat=dat)
plot_r(dat,lm)
plot_r()
}
