% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyreg_sparklyr.R
\name{plot_sparklyr}
\alias{plot_sparklyr}
\title{Plot Data and Fitted Model using sparklyr::ml_linear_regression}
\usage{
plot_sparklyr(df = gen_df_sparklyr(), fit = NULL)
}
\arguments{
\item{dat}{Polynomial Data}

\item{lm}{Linear Regression Fit on Polynomial Data}
}
\description{
Plot Data and Fitted Model using sparklyr::ml_linear_regression
}
\examples{
df <- gen_df_sparklyr(B=c(5,4,3,-2),n=4000,rng=c(-3,5),err=c(0,5))
fit <- fit_sparklyr(df=df)
plot_sparklyr(df,fit)
plot_sparklyr()
}
