% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{predict.FSR}
\alias{predict.FSR}
\title{predict.FSR}
\usage{
\method{predict}{FSR}(object, newdata, model_to_use = NULL, noisy = TRUE)
}
\arguments{
\item{object}{FSR output. Predictions will be made based on object$best_formula unless model_to_use is provided (as an integer).}

\item{newdata}{New Xdata.}

\item{model_to_use}{Integer optionally indicating a model to use if object$best_formula is not selected. Example: model_to_use = 3 will use object$models$formula[3].}
}
\value{
y_hat (predictions using chosen model estimates).
}
\description{
predict.FSR
}
