% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSR_methods.R
\name{summary.FSR}
\alias{summary.FSR}
\title{summary.FSR}
\usage{
\method{summary}{FSR}(object, estimation_overview = TRUE,
  results_overview = TRUE, model_number = NULL, ...)
}
\arguments{
\item{object}{an FSR object}

\item{estimation_overview}{logical: describe how many models were planned, sample size, etc.?}

\item{results_overview}{logical: give overview of best fit model, etc?}

\item{model_number}{If non-null, an integer indicating which model to display a summary of.}

\item{...}{ignore.}
}
\description{
summary.FSR
}
