% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{FSR}
\alias{FSR}
\title{FSR}
\usage{
FSR(Xy, max_poly_degree = 3, max_interaction_degree = 2,
  model_type = NULL, cor_type = "pearson", threshold_include = 0.01,
  threshold_estimate = 0.001, standardize = FALSE, pTraining = 0.8,
  pValidation = 0.2, min_models = NULL, max_fails = 2, file_name = NULL,
  max_block = 250, noisy = TRUE, seed = NULL)
}
\arguments{
\item{Xy}{matrix or data.frame; outcome must be in final column.}

\item{max_poly_degree}{highest power to raise continuous features; default 3 (cubic).}

\item{max_interaction_degree}{highest interaction order; default 2 (allow x_i*x_j). Also interacts each level of factors with continuous features.}

\item{model_type}{Either 'lm' (linear model), 'glm' (generalized lm), or NULL (auto-detect based on response).}

\item{cor_type}{correlation to be used for adjusted R^2; pseudo R^2 for classification. Default "pearson"; other options "spearman" and "kendall".}

\item{threshold_include}{minimum improvement to include a recently added term in the model (pseudo R^2 so scale 0 to 1). -1.001 means 'include all'. Default: 0.01.}

\item{threshold_estimate}{minimum improvement to keep estimating (pseudo R^2 so scale 0 to 1). -1.001 means 'estimate all'. Default: 0.001.}

\item{standardize}{if TRUE (not default), standardizes continuous variables.}

\item{pTraining}{portion of data for training}

\item{pValidation}{portion of data for validation}

\item{min_models}{minimum number of models to estimate. Defaults to the number of features (unless P > N).}

\item{max_fails}{maximum number of models to FSR() can fail on computationally before exiting. Default == 2.}

\item{file_name}{If a file name (and path) is provided, saves output after each model is estimated as an .RData file. ex: file_name = "results.RData"}

\item{max_block}{Most of the linear algebra is done recursively in blocks to ease memory managment. Default 250. Changing up or down may slow things...}

\item{noisy}{display measures of fit, progress, etc. Recommended.}

\item{seed}{Automatically set but can also be passed as paramater.}
}
\value{
list with slope coefficients, model details, and measures of fit
}
\description{
FSR
}
