\name{pef}
\alias{pef}

\title{
Silicon Valley programmers and engineers data
}

\usage{
data(pef)
}

\description{
This data set is adapted from the 2000 Census (5\% sample, person
records).  It is mainly restricted to programmers and engineers in the 
Silicon Valley area.  (Apparently due to errors, there are some from
other ZIP codes.)

Has columns for age, education, occupation, gender, wage income and
weeks worked.  The education column has been collapsed to Master's
degree, PhD and other.

The variable codes, e.g. occupational codes, are available from the Census
Bureau, at
\url{http://www.census.gov/prod/cen2000/doc/pums.pdf}.
(Short code lists are given in the record layout, but longer ones are in
the appendix Code Lists.)

}



