% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyreg.R
\name{polyreg}
\alias{polyreg}
\title{Direct polynomial regression for survival and competing risks analysis}
\usage{
polyreg(
  nuisance.model,
  exposure,
  strata = NULL,
  data,
  code.event1 = 1,
  code.event2 = 2,
  code.censoring = 0,
  code.exposure.ref = 0,
  effect.measure1 = "RR",
  effect.measure2 = "RR",
  time.point = NULL,
  outcome.type = "COMPETINGRISK",
  conf.level = 0.95,
  report.nuisance.parameter = FALSE,
  report.optim.convergence = FALSE,
  report.boot.conf = FALSE,
  boot.bca = TRUE,
  boot.parameter1 = 200,
  boot.parameter2 = 46,
  outer.optim.method = "nleqslv",
  inner.optim.method = "optim",
  optim.parameter1 = 1e-05,
  optim.parameter2 = 20,
  optim.parameter3 = 100,
  optim.parameter4 = 1e-05,
  optim.parameter5 = 200,
  optim.parameter6 = 1e-10,
  optim.parameter7 = 200,
  data.initial.values = NULL,
  should.normalize.covariate = TRUE,
  should.sort.data = TRUE,
  prob.bound = 1e-05
)
}
\arguments{
\item{nuisance.model}{formula Model formula representing outcome, exposure and covariates}

\item{exposure}{character Column name representing the binary exposure variable.}

\item{strata}{character Column name representing the stratification variable for adjustment for dependent censoring. Defaults to NULL.}

\item{data}{data.frame Input dataset containing the outcome, the exposure and covariates.}

\item{code.event1}{integer Specifies the code of event 1. Defaults to 1.}

\item{code.event2}{integer Specifies the code of event 2. Defaults to 2.}

\item{code.censoring}{integer Specifies the code of censoring. Defaults to 0.}

\item{code.exposure.ref}{integer Specifies the code of the reference category of exposure. Defaults to 0.}

\item{effect.measure1}{character Specifies the effect measure for event (RR, OR, SHR).}

\item{effect.measure2}{character Specifies the effect measure for competing risk (RR, OR, SHR).}

\item{time.point}{numeric The time point for exposure effects to be estimated.}

\item{outcome.type}{character Specifies the type of outcome (COMPETINGRISK, SURVIVAL, BINOMIAL and PROPORTIONAL).}

\item{conf.level}{numeric The level for confidence intervals.}

\item{report.nuisance.parameter}{logical Reports estimates and other statistics of nuisance parameters. Defaults to FALSE.}

\item{report.optim.convergence}{logical Reports indicators of convergence of parameter estimation. Defaults to FALSE.}

\item{report.boot.conf}{logical Specifies the use of bootstrap confidence intervals. Defaults to FALSE.}

\item{boot.bca}{logical Specifies the method of bootstrap confidence intervals (TRUE = BCA method, FALSE = normal approximation).}

\item{boot.parameter1}{integer Number of replicatations for bootstrap confidence intervals when outcome.type==PROPORTIONAL.}

\item{boot.parameter2}{numeric Seed used in bootstrap confidence intervals when outcome.type==PROPORTIONAL.}

\item{outer.optim.method}{character Specifies the method of optimization (nleqslv, optim, multiroot).}

\item{inner.optim.method}{character Specifies the method of optimization (nleqslv, optim, multiroot).}

\item{optim.parameter1}{numeric Convergence threshold for outer loop. Defaults to 1e-5.}

\item{optim.parameter2}{integer Maximum number of iterations. Defaults to 20.}

\item{optim.parameter3}{numeric Constraint range for parameters. Defaults to 100.}

\item{optim.parameter4}{numeric Convergence threshold for optim in outer loop. Defaults to 1e-5.}

\item{optim.parameter5}{integer Maximum number of iterations for nleqslv/optim in outer loop. Defaults to 200.}

\item{optim.parameter6}{numeric Convergence threshold for optim in inner loop. Defaults to 1e-10.}

\item{optim.parameter7}{integer Maximum number of iterations for optim in inner loop. Defaults to 200.}

\item{should.normalize.covariate}{logical Indicates whether covariates are normalized (TRUE = normalize, FALSE = otherwise). Defaults to TRUE.}

\item{should.sort.data}{logical Indicates whether data are initially sorted to reduce computation steps (TRUE = sort, FALSE = otherwise). Defaults to TRUE.}

\item{prob.bound}{numeric Small threshold for clamping probabilities. Defaults to 1e-5.}

\item{data.initlal.values}{data.frame Optional dataset containing initial values. Defaults to NULL.}
}
\value{
A list of results from direct polynomial regression. coefficient and cov are estimated regression coefficients of exposure and covariates and their variance covariance matrix. summary and summary.full meets requirement of msummary function.
}
\description{
Direct polynomial regression for survival and competing risks analysis
}
\examples{
data(diabetes.complications)
output <- polyreg(nuisance.model = Event(t,epsilon)~+1, exposure = 'fruitq1', data = diabetes.complications, effect.measure1='RR', effect.measure2='RR', time.point=8, outcome.type='C')
library(modelsummary)
msummary(output$summary, statistic = c("conf.int"), exponentiate = TRUE)
}
